refactor(ccusage): split opencode adapter sources
Separate OpenCode path discovery, schema, loader, pricing, and row aggregation so the adapter follows the shared detect/load/parse/aggregate layering.

OpenCode has multiple source kinds, so JSON message files now use the shared worker gating and SQLite remains a separate DB source. The loader still de-duplicates JSON messages against DB rows and keeps JSON+SQL totals combined.

Path checks now use the internal safe directory helper, pricing fetcher disposal uses using, and new fs-fixture tests cover path discovery, JSON message parsing, ignored non-billable messages, and row aggregation.
